implementing zoom

+++ include/fractol.h
@@ -6,7 +6,7 @@
 /*   By: vthomas <vthomas@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2016/11/09 15:13:49 by vthomas           #+#    #+#             */
-/*   Updated: 2016/11/15 21:35:45 by vthomas          ###   ########.fr       */
+/*   Updated: 2016/11/16 00:00:47 by vthomas          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -60,7 +60,7 @@ typedef struct	s_data
 	t_img		*img;
 	t_fract		*fract;
 	void		*(*draw)(void *);
-	int			zoom;
+	float		zoom;
 }				t_data;
 
 typedef struct	s_thread
